skills/vid-clip-extractor/lib/clip_extractor/requirements.txt
mediapipe>=0.10.14
opencv-contrib-python>=4.9.0
numpy>=1.24.0
pyyaml>=6.0
filterpy>=1.4.4
rapidfuzz>=3.0.0
whisperx>=3.3.0
yt-dlp>=2024.0.0
